wip: working rename to JPAFinder
