+++ jakarta-ee/flowlogix-jee/src/demo/java/com/flowlogix/demo/jeedao/UserDAO.java
@@ -16,8 +16,8 @@
 package com.flowlogix.demo.jeedao;
 
 import com.flowlogix.jeedao.DaoHelper;
-import com.flowlogix.api.dao.DaoHelper.ParameterFunction;
-import com.flowlogix.api.dao.DaoHelper.QueryEnhancement;
+import com.flowlogix.api.dao.JPAFinder.ParameterFunction;
+import com.flowlogix.api.dao.JPAFinder.QueryEnhancement;
 import com.flowlogix.demo.jeedao.entities.UserEntity;
 import com.flowlogix.demo.jeedao.entities.UserEntity_;
 import jakarta.ejb.Stateless;
