+++ jakarta-ee/flowlogix-jee/src/main/java/com/flowlogix/api/dao/JPAFinderNative.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright (C) 2011-2024 Flow Logix, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.flowlogix.api.dao;
+
+import com.flowlogix.jeedao.TypedNativeQuery;
+
+public interface JPAFinderNative<TT> extends JPAFinderHelper<TT> {
+    TypedNativeQuery createNativeQuery(String sql, Class<?> resultClass);
+    TypedNativeQuery createNativeQuery(String sql, String resultMapping);
+}
